[FEAT] Set up claim purchase
